# backend/src/main/resources/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: Anki
  description: Anki 작업
  version: 1.0.0
servers:
- url: http://localhost:8080/
  description: 8080 포트에서 사용
- url: http://localhost:7070/
  description: 7070 포트에서 사용
- url: https://virtserver.swaggerhub.com/js990311/Anki/1.0.0
  description: swagger-hub에서 사용
...
paths:
...
  /cards/{id}/study:
    get:
      tags:
      - StudyCards
      summary: Card 학습 정보 보기
      description: Card 학습 정보 보기
      operationId: user-card-controller-test/get-card-study-info
      parameters:
      - name: id
        in: path
        description: 카드의 고유번호
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: "#/components/schemas/cardStudyInfo"
              examples:
                user-card-controller-test/get-card-study-info:
                  value: "{\"cardId\":1,\"nextStudyDate\":\"2025-01-17T20:06:39.328088\"\
                    ,\"lapses\":0,\"lastReview\":null,\"reps\":0,\"scheduledDays\"\
                    :0.0,\"stability\":0.0,\"state\":\"New\"}"
...
components:
  schemas:
...
    cardStudyInfo:
      title: cardStudyInfo
      type: object
      properties:
        lapses:
          type: number
          description: Again을 누른 횟수
        reps:
          type: number
          description: 총 복습횟수
        scheduledDays:
          type: number
          description: 현재 복습 간격
        cardId:
          type: number
          description: Card의 고유번호
        lastReview:
          type: string
          description: 마지막으로 복습한 날짜
        nextStudyDate:
          type: string
          description: 다음 학습할 날짜
        state:
          type: string
          description: 카드의 현재 상태
        stability:
          type: number
          description: 기억의 안정도
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
security:
  - cookieAuth: []